api/requirements.txt
anyio==3.6.1
autoflake==1.7.1
backports.cached-property==1.0.2
black==22.8.0
click==8.1.3
fastapi==0.85.0
flake8==5.0.4
graphql-core==3.2.3
greenlet==1.1.3
h11==0.13.0
idna==3.4
importlib-metadata==4.2.0
isort==5.10.1
mccabe==0.7.0
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
Pygments==2.13.0
python-dateutil==2.8.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
strawberry-graphql==0.132.1
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.3.0
uvicorn==0.18.3
zipp==3.8.1
